Simplify Breadcrumbs structure
