+++ www/front_src/src/Breadcrumbs/models/index.ts
@@ -1,0 +1,5 @@
+export interface BreadcrumbItem {
+  link: string;
+  label: string;
+  index: number;
+}
